+++ hotness/buildsys.py
@@ -315,7 +315,7 @@ class Koji(object):
         """
 
         # Clone the package to a tempdir
-        tmp = tempfile.mkdtemp(prefix="thn-")
+        tmp = tempfile.mkdtemp(prefix="thn-", dir="/var/tmp")
         try:
             url = self.git_url.format(package=package)
             _log.info("Cloning %r to %r" % (url, tmp))
@@ -446,7 +446,7 @@ def dist_git_sources(dist_git_path):
     # ####################################################################### 100.0%
     # Downloading requests-2.12.4-tests.tar.gz
     # ####################################################################### 100.0%
-    output = sp.check_output(["fedpkg", "sources"], cwd=dist_git_path)
+    output = sp.check_output(["fedpkg", "--user hotness", "sources"], cwd=dist_git_path)
     for line in output.splitlines():
         if line.startswith("Downloading"):
             files.append(os.path.join(dist_git_path, line.split()[-1]))