+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-master
+0.11.4
 ------
 
 Bugfixes